+++ src/module/setupModules.ts
@@ -88,6 +88,18 @@ export function checkCubInstalled() {
 Hooks.once('libChangelogsReady', function() {
   //@ts-ignore
   libChangelogs.register("midi-qol",`
+0.8.73
+* A little tidying of active defence rolls so that duplicate rolls are not performed.
+* Fix for midi-qol.RollComplete firing too early in the workflow, leaving workflow.damageList undefined.
+* Added fumbleSaves/criticalSaves: Set<Token> to workflow, and fumbleSaves,criticalSaves,fumbleSaveUuids, criticlSaveUuids to onUse/damageBonus macro arguments.
+
+0.8.72
+* Fix for active defence error in ac defence roll calculation.
+* Added support for ItemMacro.UUID in DamageBonusMacros and OnUse macros to refernce item macros for items not in your inventory.
+
+0.8.71
+* Fix for active defence causing a console error for non gm clients.
+
 0.8.70  
   * Fix for damage type none and better rolls (would always do 0 damage).
   * Expirmental: Support for the Active Defence variant rule. Enable via optional rules setting Active Defence. Requires LIMRTFY and does **not** work with better rolls. 
@@ -101,36 +113,12 @@ Hooks.once('libChangelogsReady', function() {
     - Issues: There is only one critical result supported, so if multiple targets are attacked they will all have critical damage rolled against them or none. (future might support individual results)
     - There is only 1 advantaage/disadvantage setting applied, that of the first defender (same as current midi-qol). Future enhancement will use per character advantage/disadvantage settings.
     - Only works for mwak/rwak/rsak/msak.
-    
+
 0.8.69
 **Changes coming in dnd5e 1.5**:
 * dnd5e 1.5 includes per weapon critical threshold and bonus critical damage dice. There is now a configuration setting to enable/disable the midi-qol field on the item sheet. Once dnd5e 1.5 is released, you are stongly encouraged to migrate to the dnd5e setting and disable the midi-qol flag, via Use Midi Critical in the configuration settings. Soon, I will remove the midi-qol field completely. 
 * You can run MidiQOL.reportMidiCriticalFlags() from the console to see which actors/tokens have the midi-qol critical setting defined.
 * Enhanced dnd5e critical damage effects. You can make most of the changes that midi-qol supports for critical hits via the new game settings (max base dice, double modifiers as well as dice) and per weapon settings (additional dice). You will need to experiment to cofirm the interaction of the dnd5e critical damage flags and the midi-qol settings, however if you use the dnd5e default setting in midi-qol the rolls will not be modified by midi in any way and the dnd5e system will operate.
-
-0.8.68
-* Fix for betterrolls and OverTime effects not rolling damage correctly/at all.
-* Fix for bettereolls saving throws not being detected in chat message saves workflow.
-* Overtime effects saveDC now supports expressions rather than just numbers/field lookups. No dice expressions.
-* Fix for reaction checks throwing an error if no midi-qol flags are set on the actor.
-
-0.8.65/66/67
-* Fixes for template targeting and various module interactions. (Including the elusive bug exposed when token magic fx not installed).
-
-0.8.64
-* Added healing and temp helaing to resistance/immunity/vulnerability types so that actor can be immune to healing.
-
-0.8.63
-* Fix for OverTime effects - now supports stacking of the same effect (should be regarded as experimental).
-* Added the ability to use # instead of , to separate OverTime fields. You have to use one or the other for the whole OverTime field.
-* Fix for BonusDamageRoll rolls and dice so nice not being displayed.
-* new MidiQOL.completeItemRoll function. Can use with  await MidiQOL.completeItemRoll(ownedItem, options) which will return the worfklow when the entire roll is complete and damage (if any) applied.
-* Template targeting clean up. 
-  - If using levels you can set the midi-qol optional rule setting to check + levels, which will check template overage including height and levels walls blocking for all attacks. The midi template height check is VERY naive and in addition to 2d targeting simply checks a sphere centered on the template origin and if further away it is considered out of range.
-  - If you want proper volumetic templates use the levelsvolumetrictemplates module (patreon) which does a great job of working out how much of the token is in the template and midi uses the result of that calculation. This version supports walls blocking for volumetric templates and uses volumetric templates for the preview targeting.
-  - Midi also supports the levels "place next template at this height" setting from the left hand hud and if not set, will cast at the tokens current LoS height.
-  - If levels is installed midi will use levels' check collision code, which deals with wall heights.
-
   
   [Full Changelog](https://gitlab.com/tposney/midi-qol/-/blob/master/Changelog.md)`,
   "major")